Removed remaining parts of Geomeotries.h/cpp
