+++ client/gui/TextAlignment.h
@@ -1,0 +1,12 @@
+/*
+ * TextAlignment.h, part of VCMI engine
+ *
+ * Authors: listed in file AUTHORS in main folder
+ *
+ * License: GNU General Public License v2.0 or later
+ * Full text of license available in license.txt file, in main folder
+ *
+ */
+#pragma once
+
+enum class ETextAlignment {TOPLEFT, CENTER, BOTTOMRIGHT};
